requirements.txt
aiogram==3.22.0 
redis==6.4.0
SQLAlchemy==2.0.43
sshtunnel==0.4.0
asyncpg==0.30.0 
python-decouple
APScheduler==3.11.0 
paramiko<3.0.0
